use a Queue for failed read 'report'
